pyproject.toml
[tool.poetry]
name = "final-project-be"
version = "0.0.1"
description = "Healthy Recipe"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.0.3"
flask-sqlalchemy = "^3.1.1"
flask-smorest = "^0.44.0"
docopt = "^0.6.2"
cerberus = "^1.3.5"
google-auth = "^2.29.0"
idna = "^3.6"
python-dotenv = "^1.0.1"
rsa = "^4.9"
flask-jwt-extended = "^4.6.0"
flask-login = "^0.6.3"
bcrypt = "^4.1.2"
blinker = "^1.7.0"
cachetools = "^5.3.3"
certifi = "^2024.2.2"
charset-normalizer = "^3.3.2"
click = "^8.1.7"
jinja2 = "^3.1.3"
itsdangerous = "^2.1.2"
markupsafe = "^2.1.5"
mysql-connector-python = "^8.3.0"
pyasn1 = "^0.6.0"
pyasn1-modules = "^0.4.0"
pyjwt = "^2.8.0"
requests = "^2.31.0"
sqlalchemy = "^2.0.29"
typing-extensions = "^4.11.0"
urllib3 = "^2.2.1"
werkzeug = "^3.0.2"
yarg = "^0.1.9"
flask-migrate = "^4.0.7"
psycopg2 = "^2.9.9"
passlib = "^1.7.4"
supabase = "^2.4.3"
pytest = "^8.1.1"
sentry-sdk = {extras = ["flask"], version = "^1.45.0"}
marshmallow-enum = "^1.5.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
